verloop/requirements.txt
amqp==2.5.1
anyjson==0.3.3
arrow==0.15.7
asgiref==3.5.0
asn1crypto==0.24.0
attrs==16.3.0
auditlog3==1.0.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
billiard==3.6.0.0
celery==4.4.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
click-completion==0.5.1
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
defusedxml==0.7.1
Django==3.2.11
django-debug-toolbar==2.1
django-debug-toolbar-request-history==0.1.0
django-filter==2.1.0
django-guardian==2.0.0
django-redis==4.11.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-xml==2.0.0
docutils==0.15.2
drf-nested-routers==0.91
geographiclib==1.49
idna==2.8
imagesize==1.1.0
importlib-metadata==0.19
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
jwt==0.6.1
kombu==4.6.4
ladon==1.0.4
lxml==4.5.1
Markdown==3.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
novo-jsobj==0.0.2
openapi-codec==1.3.2
optionaldict==0.1.1
packaging==19.1
paramiko==2.7.2
passlib==1.7.1
pdfkit==0.6.1
Pillow==6.1.0
psycopg2==2.8.5
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
PyNaCl==1.4.0
pyodbc==4.0.27
pyparsing==2.4.2
python-dateutil==2.8.0
python-memcached==1.59
pytz==2019.1
PyYAML==5.1.1
qrcode==6.1
redis==3.3.6
requests==2.23.0
requests-toolbelt==0.9.1
rjsmin==1.1.0
shellescape==3.4.1
shellingham==1.3.1
simplejson==3.16.0
six==1.12.0
snowballstemmer==1.9.0
soupsieve==2.3.1
split-settings==1.0.0
spyne==2.12.16
SQLAlchemy==1.2.7
sqlparse==0.3.0
tornado==5.1.1
typing-extensions==4.0.1
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
Werkzeug==0.15.5
xlrd==1.2.0
XlsxWriter==1.2.8
xmltodict==0.12.0
zipp==0.5.2